Add More Points, Tests, Fix Bug
I didn't find the end gate.

I thought my math was right, so I collected more data. My results really
didn't look right but I couldn't find the error. So I added tests to
prove things were or weren't working. Found a bug in the math. I needed
to flip the y-intercepts in the Line.intersect method. Problem solved.
Getting results that look right.

Line cd also looked way off, I think because the line was too short, and minecraft is
really blocky, and eye of ender drifts a bit between stopping in the air
and landing on the ground, and I think point d was where the eye of ender
landed, giving a less than accurate slope.

So it was a combination of the math and the data.